+++ src/main/java/com/mraof/minestuck/data/MSAdvancementProvider.java
@@ -62,6 +62,7 @@ public class MSAdvancementProvider implements ForgeAdvancementProvider.Advanceme
 	public static final String BIG_ONE_MIL = "minestuck.big_one_mil";
 	public static final String INTELLIBEAM = "minestuck.intellibeam";
 	public static final String LEGENDARY_WEAPON = "minestuck.legendary_weapon";
+	public static final String BUY_OUT_SHOP = "minestuck.buy_out_shop";
 	
 	public static DataProvider create(PackOutput output, CompletableFuture<HolderLookup.Provider> registries, ExistingFileHelper existingFileHelper)
 	{
@@ -96,6 +97,7 @@ public class MSAdvancementProvider implements ForgeAdvancementProvider.Advanceme
 		Advancement bigOneMil = Advancement.Builder.advancement().parent(halfwayPoint).display(BoondollarsItem.setCount(new ItemStack(MSItems.BOONDOLLARS.get()), 1_000_000), Component.translatable(title(BIG_ONE_MIL)), Component.translatable(desc(BIG_ONE_MIL)), null, FrameType.GOAL, true, true, false).addCriterion("reach_rung", EcheladderTrigger.Instance.rung(MinMaxBounds.Ints.atLeast(44))).save(saver, save_loc(BIG_ONE_MIL));
 		Advancement intellibeam = Advancement.Builder.advancement().parent(alchemy).display(MSItems.INTELLIBEAM_LASERSTATION.get(), Component.translatable(title(INTELLIBEAM)), Component.translatable(desc(INTELLIBEAM)), null, FrameType.TASK, true, true, false).addCriterion("use_intellibeam", IntellibeamLaserstationTrigger.Instance.any()).save(saver, save_loc(INTELLIBEAM));
 		Advancement strongWeapon = Advancement.Builder.advancement().parent(alchemy).display(MSItems.CALEDFWLCH.get(), Component.translatable(title(LEGENDARY_WEAPON)), Component.translatable(desc(LEGENDARY_WEAPON)), null, FrameType.TASK, true, true, false).addCriterion("get_max_tier_weapon", InventoryChangeTrigger.TriggerInstance.hasItems(ItemPredicate.Builder.item().of(MSTags.Items.LEGENDARY).build())).save(saver, save_loc(LEGENDARY_WEAPON));
+		Advancement buyOutShop = Advancement.Builder.advancement().parent(commune).display(MSItems.CONE_OF_FLIES.get(), Component.translatable(title(BUY_OUT_SHOP)), Component.translatable(desc(BUY_OUT_SHOP)), null, FrameType.TASK, true, true, false).addCriterion("buy_everything", EventTrigger.Instance.buyOutShop()).save(saver, save_loc(BUY_OUT_SHOP));
 	}
 	
 	private static Advancement.Builder changeModusCriteria(Advancement.Builder builder)